% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/blockspringDefine.R
\name{blockspringDefine}
\alias{blockspringDefine}
\title{This function defines a new blockspring function.}
\usage{
blockspringDefine(block)
}
\arguments{
\item{block}{A function.}
}
\value{
This function doesn't return anything, it strictly defines a new blockspring function.
}
\description{
This function defines a new blockspring function.
}

