% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/blockspringParse.R
\name{blockspringParse}
\alias{blockspringParse}
\title{This function parses a Blockspring response.}
\usage{
blockspringParse(input_params, json_parsed)
}
\arguments{
\item{input_params}{A list.}

\item{json_parsed}{A boolean.}
}
\value{
This function parses and returns the result of a Blockspring response.
}
\description{
This function parses and returns the result of a Blockspring response.
}

