% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/blockspringRunParsed.R
\name{blockspringRunParsed}
\alias{blockspringRunParsed}
\title{This function allows you to run code remotely and get back parsed result.}
\usage{
blockspringRunParsed(block, data)
}
\arguments{
\item{block}{A string.}

\item{data}{A list.}

\item{api_key}{A string.}
}
\value{
This function returns the result of running a blockspring block.
}
\description{
This function allows you to run code remotely.
}

