% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/define.R
\name{define}
\alias{define}
\title{This function defines a new blockspring function.}
\usage{
define(block)
}
\arguments{
\item{block}{A function.}
}
\value{
This function doesn't return anything, it strictly defines a new blockspring function.
}
\description{
This function defines a new blockspring function.
}
\examples{
define(myFunction)
}

