% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{This function allows you to run code remotely.}
\usage{
run(block, data)
}
\arguments{
\item{block}{A string.}

\item{data}{A list.}
}
\value{
This function doesn't return anything, it strictly defines a new blockspring function.
}
\description{
This function allows you to run code remotely.
}
\examples{
run("/blockspring_username/block_id", list())
}

